Rewrites the btsnoop parser to be more Python 3 friendly
The rewrite is to me more readable in the sense that there are not
`assert` statements incorporated into the parsing, nor are there
any print statements (at least, not anymore? debug by print is a
useful tool, but not needed anymore, for now). Instead, the valiation
is incorporated into the parse result, by way of an error message.
This provides the ability to display the error, without assuming a
console/terminal session is the current user interface.

WHERE ARE THE TESTS!?!?!?!

Yeah, about them.

I'll add them soon (tm). I write my code in such a way as to promote
testing, but I don't write tests for unwritten code, due to the nature
of not knowing all the code paths that I would be testing. But tests,
they will be added!

HCI_CMD/EVT's

Yeah, they'll arrive. Eventually. I'm more interested in understanding
the device communication protocol. If I get to it, I might even add
those commands/events too, as they don't seem to be very well documented,
but that's gonna be some reverse engineering work, for sure!

Finally, poke some fun at VIM too. Because why not?